chore: scope curtailment modal branch
