+++ client/src/protoFleet/features/settings/components/Schedules/ScheduleModal.tsx
@@ -34,8 +34,6 @@ import {
   timeOptions,
   validateSchedule,
 } from "@/protoFleet/features/settings/components/Schedules/scheduleValidation";
-import TargetSelectButton from "@/protoFleet/features/settings/components/Schedules/TargetSelectButton";
-import { getTargetButtonLabel } from "@/protoFleet/features/settings/components/Schedules/targetSelectButtonLabels";
 import {
   formatDateValue,
   parseDate as parseScheduleDate,
@@ -77,6 +75,37 @@ interface ScheduleModalProps {
 const sectionTitleClassName = "text-emphasis-300 text-text-primary";
 const sectionBodyClassName = "grid gap-4";
 const popoverViewportPadding = minimalMargin * 2;
+
+const getTargetButtonLabel = (count: number, singular: string) =>
+  count > 0 ? `${count} ${count === 1 ? singular : `${singular}s`}` : "Select";
+
+interface TargetSelectButtonProps {
+  label: string;
+  value: string;
+  onClick: () => void;
+}
+
+const TargetSelectButton = ({ label, value, onClick }: TargetSelectButtonProps) => {
+  const isPlaceholder = value === "Select";
+
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className="relative flex h-14 w-full items-center justify-between rounded-lg border border-border-5 bg-surface-base pr-4 pl-4 text-left outline-hidden"
+    >
+      <div className="flex min-w-0 flex-col pt-[18px]">
+        <span className="absolute top-[7px] text-200 text-text-primary-50">{label}</span>
+        <div
+          className={isPlaceholder ? "truncate text-300 text-text-primary-50" : "truncate text-300 text-text-primary"}
+        >
+          {value}
+        </div>
+      </div>
+      <ChevronDown width="w-3" className="shrink-0 text-text-primary-70" />
+    </button>
+  );
+};
 
 const weekdayMenuOptions: Array<{ value: DayOfWeek; label: string; shortLabel: string }> = [
   { value: DayOfWeek.SUNDAY, label: "Sunday", shortLabel: "Sun" },
